Docstring explanation of _is_compatible()
